% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.component.R
\docType{data}
\name{AppComponent}
\alias{AppComponent}
\title{Web Component Application}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppComponent
}
\description{
This is abstract class to create a web component,
it has five main methods \code{\link[=AppComponent$view]{AppComponent$view()}}, \code{\link[=AppComponent$server]{AppComponent$server()}}
\code{\link[=AppComponent$route]{AppComponent$route()}}, \code{\link[=AppComponent$title]{AppComponent$title()}} and \code{\link[=AppComponent$namespaceID]{AppComponent$namespaceID()}}.
}
\details{
\code{view()}:

\code{server()}:

\code{route()}:
}
\keyword{datasets}
