% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.module.R
\docType{data}
\name{AppModule}
\alias{AppModule}
\title{Web Module Application}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppModule
}
\description{
This is abstract class to create a web module,
it has three main methods \code{\link[=AppModule$sidebar]{AppModule$sidebar()}}, \code{\link[=AppModule$routes]{AppModule$routes()}}
and \code{\link[=AppModule$servers]{AppModule$servers()}}
}
\details{
\code{view()}:

\code{server()}:

\code{route()}:
}
\keyword{datasets}
