% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.root.R
\docType{data}
\name{AppRoot}
\alias{AppRoot}
\title{Application Root Container}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppRoot
}
\description{
This class is the main web container which contains of sub-modules,
it has four main methods \code{\link[=AppRoot$initialize]{AppRoot$initialize()}}, \code{\link[=AppRoot$server]{AppRoot$server()}}
\code{\link[=AppRoot$ui]{AppRoot$ui()}} and \code{\link[=AppRoot$addModule]{AppRoot$addModule()}}.
}
\details{
\code{ui()}: it create the main web interface with routes of
sub-modules injected.

\code{server()}: it create the main server (container of servers)
in order to responds all http request.

\code{addModule(module)}: add new modules to container main web,
param module is a class module that has server, siderbars and routes
}
\keyword{datasets}
