% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.scriptR}
\alias{read.scriptR}
\title{Read a script R and save in a variable}
\usage{
read.scriptR(fpath)
}
\arguments{
\item{fpath}{path file to read}
}
\description{
Read a script R and save in a variable
}
