% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.root.R
\docType{data}
\name{ServerRoot}
\alias{ServerRoot}
\title{Server Root Routes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ServerRoot
}
\description{
This isclass save the namespace of all request servers (i.e. routes/servlet).
}
\details{
\code{set()}:

\code{get()}:
}
\keyword{datasets}
