% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.module.R
\docType{data}
\name{SimpleModule}
\alias{SimpleModule}
\title{Simple Implementation Web Module Application}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SimpleModule
}
\description{
This is default simple implementation of AppModule class that create a web module component,
it has three main methods \code{\link[=AppModule$sidebar]{AppModule$sidebar()}}, \code{\link[=AppModule$routes]{AppModule$routes()}}
and \code{\link[=AppModule$servers]{AppModule$servers()}}
}
\details{
\code{view()}:

\code{server()}:

\code{route()}:
}
\keyword{datasets}
