% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tagAssert}
\alias{tagAssert}
\title{Method for checking the inheritance of a component tag}
\usage{
tagAssert(tag)
}
\arguments{
\item{tag}{variable}
}
\description{
Method for checking the inheritance of a component tag
}
