% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.module.R
\docType{data}
\name{ViewFactory}
\alias{ViewFactory}
\title{View Factory Life Cycle}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ViewFactory
}
\description{
This is static class.
}
\details{
\code{createContent()}:
}
\keyword{datasets}
