% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waheco.R
\docType{package}
\name{WAHEco}
\alias{WAHEco}
\alias{WAHEco-package}
\title{WAHEco: Web Application for Health Economics Decision Making}
\description{
\code{WAHEco} is an R toolset for creating a general custom
web application. It has a demo web modules for medico-economics
decision making, and also, anyone can create new customized modules
or components for later they are added to the web application
container AppRoot.
}
