% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeMEco.R
\docType{package}
\name{WeMeco}
\alias{WeMeco}
\alias{WeMeco-package}
\title{WeMEco: Web Application for Medico-Economics Decision Making}
\description{
\code{WeMEco} is an R toolset for creating a general custom
web application. It has a demo web modules for medico-economics
decision making, and also, anyone can create new customized modules
or components for later they are added to the web application
container AppRoot.
}
